feat(config-api): audit log, agama ADS spec, fix for 0 index search
