+++ jans-config-api/profiles/local/test.properties
@@ -4,6 +4,6 @@ test.scopes=https://jans.io/oauth/config/acrs.readonly https://jans.io/oauth/con
 # jans.server
 token.endpoint=https://jans.server1/jans-auth/restv1/token
 token.grant.type=client_credentials
-test.client.id=1800.5957dfad-b2cb-4764-85fe-841e6bc870ff
-test.client.secret=ozu4fjIzoEbe
+test.client.id=1800.c94f1e10-7716-4dc8-b82d-4dd1169ed4f9
+test.client.secret=2M6r3vYeQEIT
 test.issuer=https://jans.server1/
