+++ jans-config-api/shared/src/main/java/io/jans/configapi/core/interceptor/RequestAuditInterceptor.java
@@ -16,5 +16,5 @@ import jakarta.interceptor.InterceptorBinding;
 @InterceptorBinding
 @Retention(RetentionPolicy.RUNTIME)
 @Target({ ElementType.TYPE, ElementType.METHOD })
-public @interface RequestInterceptor {
+public @interface RequestAuditInterceptor {
 }
